+++ src/core/engine/search/Variation.h
@@ -8,9 +8,9 @@
 
 struct Variation {
     std::vector<Move> moves;
-    int32_t score;
-    uint16_t depth;
-    uint16_t selectiveDepth;
+    int score;
+    int depth;
+    int selectiveDepth;
 };
 
 template<>
